+++ src/text.ts
@@ -78,7 +78,7 @@
  */
 
 import { isObject } from "@metreeca/core";
-import { immutable } from "@metreeca/core/deep";
+import { immutable } from "@metreeca/core/structures";
 import type { Tag, TagRange } from "@metreeca/core/language";
 import type { Locale } from "@metreeca/qest/template";