+++ src/union.ts
@@ -62,7 +62,7 @@
  */
 
 import type { Lazy } from "@metreeca/core";
-import { immutable } from "@metreeca/core/deep";
+import { immutable } from "@metreeca/core/structures";
 import { eager, type ValueShape } from "./value.js";
 
 export { getShapeVariants, getStateVariant, getBoundVariant, getModelVariants } from "./union.core.js";